refactored dark mode toggle into seperate component + cleaned up some unused things
