+++ src/app/(home)/Drawer.js
@@ -1,8 +1,7 @@
-
 // React
-import React, { useState } from 'react';
+import React, {useState} from 'react';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation'
+import {usePathname} from 'next/navigation'
 
 // Material UI
 import Drawer from '@mui/material/SwipeableDrawer';
@@ -15,7 +14,7 @@ import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
 import Collapse from '@mui/material/Collapse';
-import { useTheme } from '@mui/material/styles';
+import {useTheme} from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 
 // Material UI: Icons
@@ -42,36 +41,37 @@ import ExpandMore from '@mui/icons-material/ExpandMore';
  * @param {Function} props.handleDrawerOpen - The function to handle the opening of the drawer.
  *
  * @returns {ReactElement} The React Element created by this function.
- */export default function DrawerComponent({ drawerWidth, open, handleDrawerClose, handleDrawerOpen }) {
+ */export default function DrawerComponent({drawerWidth, open, handleDrawerClose, handleDrawerOpen}) {
 
     // NextJS: Get current URL pathname
     const pathname = usePathname()
+
 
     // React Vars and Hooks
     const theme = useTheme();
     const isMobile = useMediaQuery(theme.breakpoints.down('md'));   // Using a media query to determine screen size
     const drawerVariant = isMobile ? 'temporary' : 'permanent'; // Determine drawer variant based on screen size
-    const [pathwaysOpen, setPathwaysOpen] = useState(true);
-    const [tasksOpen, setTasksOpen] = useState(true);
-
-    const [settingsOpen, setSettingsOpen] = useState(true);
-    const [supportOpen, setSupportOpen] = useState(true);
-  
-    const handlePathwaysClick = () => {
-        setPathwaysOpen(!pathwaysOpen);
-    };
-
-    const handleTasksClick = () => {
-        setTasksOpen(!tasksOpen);
-    };
-
-    const handleSettingsClick = () => {
-        setSettingsOpen(!settingsOpen);
-    };
-
-    const handleSupportClick = () => {
-        setSupportOpen(!supportOpen);
-    };
+    // const [pathwaysOpen, setPathwaysOpen] = useState(true);
+    // const [tasksOpen, setTasksOpen] = useState(true);
+    //
+    // const [settingsOpen, setSettingsOpen] = useState(true);
+    // const [supportOpen, setSupportOpen] = useState(true);
+    //
+    // const handlePathwaysClick = () => {
+    //     setPathwaysOpen(!pathwaysOpen);
+    // };
+    //
+    // const handleTasksClick = () => {
+    //     setTasksOpen(!tasksOpen);
+    // };
+    //
+    // const handleSettingsClick = () => {
+    //     setSettingsOpen(!settingsOpen);
+    // };
+    //
+    // const handleSupportClick = () => {
+    //     setSupportOpen(!supportOpen);
+    // };
 
     return (
         <Drawer
@@ -110,19 +110,19 @@ import ExpandMore from '@mui/icons-material/ExpandMore';
                 }),
             }}
         >
-            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'flex-end', ...theme.mixins.toolbar }}>
+            <Box sx={{display: 'flex', alignItems: 'center', justifyContent: 'flex-end', ...theme.mixins.toolbar}}>
                 <IconButton onClick={handleDrawerClose}>
-                    {theme.direction === 'rtl' ? <ChevronRightIcon /> : <ChevronLeftIcon />}
+                    {theme.direction === 'rtl' ? <ChevronRightIcon/> : <ChevronLeftIcon/>}
                 </IconButton>
             </Box>
             <List>
 
                 {/* Pathways */}
-                <ListItemButton onClick={handlePathwaysClick} href='/pathways' selected={pathname === '/pathways'}>
+                <ListItemButton href='/pathways' selected={pathname === '/pathways'}>
                     <ListItemIcon>
-                        <PathIcon />
+                        <PathIcon/>
                     </ListItemIcon>
-                    <ListItemText primary="Pathways" />
+                    <ListItemText primary="Pathways"/>
                     {/* {pathwaysOpen ? <ExpandLess /> : <ExpandMore />} */}
                 </ListItemButton>
                 {/* <Collapse in={pathwaysOpen} timeout="auto" unmountOnExit>
@@ -130,11 +130,11 @@ import ExpandMore from '@mui/icons-material/ExpandMore';
                 </Collapse> */}
 
                 {/* Tasks */}
-                <ListItemButton onClick={handleTasksClick} href='/tasks' selected={pathname === '/tasks'}>
+                <ListItemButton href='/tasks' selected={pathname === '/tasks'}>
                     <ListItemIcon>
-                        <TaskIcon />
+                        <TaskIcon/>
                     </ListItemIcon>
-                    <ListItemText primary="Tasks" />
+                    <ListItemText primary="Tasks"/>
                     {/* {tasksOpen ? <ExpandLess /> : <ExpandMore />} */}
                 </ListItemButton>
                 {/* <Collapse in={tasksOpen} timeout="auto" unmountOnExit>
@@ -144,27 +144,27 @@ import ExpandMore from '@mui/icons-material/ExpandMore';
             </List>
 
             {/* Spacer element to push the following List to the bottom */}
-            <Box sx={{ flexGrow: 1 }} />
-            <Divider />
+            <Box sx={{flexGrow: 1}}/>
+            <Divider/>
 
             <List>
-                <ListItemButton onClick={handleSupportClick} href='/support' selected={pathname === '/support'}>
+                <ListItemButton href={"/support"} selected={pathname === '/support'}>
                     <ListItemIcon>
-                        <SupportIcon />
+                        <SupportIcon/>
                     </ListItemIcon>
-                    <ListItemText primary="Support" />
+                    <ListItemText primary="Support"/>
                 </ListItemButton>
-                <ListItemButton>
+                <ListItemButton href="/home">
                     <ListItemIcon>
-                        <HomeIcon />
+                        <HomeIcon/>
                     </ListItemIcon>
-                    <ListItemText primary="Home Page" href='/#' />
+                    <ListItemText primary="Home Page"/>
                 </ListItemButton>
-                <ListItemButton onClick={handleSettingsClick} href='/settings' selected={pathname === '/settings'}>
+                <ListItemButton href={"/settings"} selected={pathname === '/settings'}>
                     <ListItemIcon>
-                        <SettingsIcon />
+                        <SettingsIcon/>
                     </ListItemIcon>
-                    <ListItemText primary="Settings" />
+                    <ListItemText primary="Settings"/>
                 </ListItemButton>
             </List>
         </Drawer>
